+++ src/ts/interfaces.ts
@@ -49,3 +49,12 @@ export interface IDialogueGraph<T extends IDialogueNodeType> extends Record<TDia
 export interface IDialogueConfig<T extends IDialogueNodeType> {
   graph: IDialogueGraph<T>;
 }
+
+/**
+ * Used to identify options.
+ */
+export interface IOptionIdentifyer {
+  nodeName: string;
+  node: IMiniQuestionaireWithPosition;
+  optionIndex: number;
+}