+++ changelog.md
@@ -2,6 +2,7 @@
 
 - 2023-08-03
   - Adding highlighted options connectors (on mouse over).
+  - Can visually reconnect nodes now (by mouse).
 - 2023-08-02
   - Fixing drag amount when moving boxes.
   - Drawing connections more beautiful with cubic Bézier curves.